+++ scripts/build.js
@@ -2,13 +2,15 @@
  * @Author: junjie.lean
  * @Date: 2020-06-17 16:51:24
  * @Last Modified by: junjie.lean
- * @Last Modified time: 2020-07-29 15:56:09
+ * @Last Modified time: 2021-04-21 13:50:07
  */
 
 const spawn = require("cross-spawn");
 const zipper = require("zip-local");
 const chalk = require("chalk");
 const package = require("./../package");
+const fs = require("fs");
+
 process.on("unhandledRejection", (err) => {
   throw err;
 });
@@ -33,7 +35,6 @@ let res = spawn.sync(
   }
 );
 
-
 console.log(chalk.hex("#00d684").bold("Zip start, please wait..."));
 
 let hash = "";
@@ -45,8 +46,9 @@ let zipName = `build${
   package.projectName ? "." + package.projectName : ""
 }.${hash}.zip`;
 
-zipper.sync.zip("./build").compress().save(zipName);
-
-console.log(
-  chalk.hex("#00d684").bold(`Zip completed, package name: "${zipName}"`)
-);
+if (fs.existsSync("./build")) {
+  zipper.sync.zip("./build").compress().save(zipName);
+  console.log(
+    chalk.hex("#00d684").bold(`Zip completed, package name: "${zipName}"`)
+  );
+}