+++ src/view/page/layout-home.js
@@ -2,50 +2,69 @@
  * @Author: junjie.lean
  * @Date: 2020-03-18 11:00:47
  * @Last Modified by: junjie.lean
- * @Last Modified time: 2020-11-16 10:57:09
+ * @Last Modified time: 2021-04-21 12:24:33
  */
 
 import React, { useEffect, useState, useRef } from "react";
 import "./../../style/index.scss";
-import { Button } from "antd";
+import { Button, Input } from "antd";
 import { useSelector } from "react-redux";
-
+import Axios from "axios";
 function Home(props) {
-  const [stringSource] = useState("  Write nothing, deploy nowhere...");
-  const [string, setStr] = useState("");
+  const [count, setCount] = useState(1);
+  const [res, setRes] = useState();
+  const [isReady, setReadyState] = useState(false);
+  const inputRef = useRef();
 
-  let timer = useRef();
-
-  useEffect(() => {
-    let b = 0;
-    timer.current = setInterval(() => {
-      setStr(stringSource.slice(0, b + 1));
-      b++;
-      if (b == stringSource.length) {
-        clearInterval(timer.current);
-      }
-    }, 150);
-    return () => {
-      clearInterval(timer.current);
-    };
-  }, []);
-
-  let reduxProps = useSelector((state) => state);
+  const getData = (page) => {
+    Axios.get("https://www.randomuser.me/api?page=" + page).then((res) => {
+      setRes(res.data);
+      setReadyState(true);
+    });
+  };
 
   useEffect(() => {
-    console.log(reduxProps);
-  }, []);
+    getData(count);
+  }, [count]);
 
   return (
-    <div className="lean-homeStyle">
-      <div>
-        <Button type="link" style={{ paddingRight: 2, fontSize: 32 }}>
-          <a href="https://github.com/kelseyhightower/nocode" target="_blank">
-            {string}
-          </a>
-        </Button>
-        <span className="lean-homeCursor"> | </span>
-      </div>
+    <div style={{ padding: 20 }}>
+      <p>---------- {count}</p>
+      <Button
+        type="primary"
+        onClick={() => {
+          setCount(count + 1);
+        }}
+      >
+        计数器+1
+      </Button>
+      <br />
+      <br />
+      <br />
+      {/* <Button
+        type="primary"
+        onClick={() => {
+          setCount(count - 1);
+        }}
+      >
+        计数器-1
+      </Button> */}
+      <br />
+      <br />
+      {isReady ? (
+        <div>
+          {res?.results?.map((item) => {
+            return (
+              <div key={item.email}>
+                <div>name:{item.name?.first}</div>
+                <div>gender:{item.gender}</div>
+              </div>
+            );
+          })}
+        </div>
+      ) : (
+        "请求中.."
+      )}
     </div>
   );
 }